+++ CHANGELOG.md
@@ -36,8 +36,10 @@ All notable changes to this project are documented in this file.
 
 ### Fixed
 
+- Code
+    - ``awsmate.apigateway``: corrected type hints for ``**kwargs``
 - Example application
-    - Deployment scripts: ``.`` was mistakenly assumed to be in the ``PATH``.
+    - deployment scripts: ``.`` was mistakenly assumed to be in the ``PATH``.
     - API Gateway: fixed the redeployment issue
 
 ### Security